+++ packages/plugin-ext/src/common/plugin-protocol.ts
@@ -442,6 +442,11 @@ export interface PluginModel {
         version: string;
     };
     entryPoint: PluginEntryPoint;
+    packageUri: string;
+    /**
+     * @deprecated since 1.1.0 - because it lead to problems with getting a relative path
+     * needed by Icon Themes to correcty load Fonts, use packageUri instead.
+     */
     packagePath: string;
     iconUrl?: string;
     readmeUrl?: string;